requirements.txt
Django
djangorestframework
mysqlclient
corsheaders
uwsgi